refactor: Field 컴포넌트의 위치를 layout에서 common으로 변경
