+++ src/components/ExpirationDateInput/ExpirationDateInput.tsx
@@ -1,5 +1,5 @@
 import Input from '../common/Input/Input';
-import Field from '../layout/Field/Field';
+import Field from '../common/Field/Field';
 
 import {
   isInteger,
